refactoring
https://github.com/metasfresh/metasfresh/issues/11560
