+++ backend/de.metas.handlingunits.base/src/main/java/de/metas/handlingunits/attribute/IHUAttributesBL.java
@@ -5,7 +5,6 @@ import de.metas.handlingunits.model.I_M_HU;
 import de.metas.util.ISingletonService;
 import lombok.NonNull;
 import org.adempiere.mm.attributes.AttributeCode;
-import org.adempiere.mm.attributes.AttributeId;
 import org.adempiere.mm.attributes.api.IAttributeSet;
 import org.compiere.model.I_M_Attribute;
 
@@ -50,7 +49,7 @@ public interface IHUAttributesBL extends ISingletonService
 
 	/**
 	 * Iterates the HU-tree of the given HU and sets the given attribute to the given attributeValue.
-	 *
+	 * <p>
 	 * Note: for complex scenarios (distributing a weight onto an HU-tree), see {@link de.metas.handlingunits.attribute.propagation.IHUAttributePropagator} and {@link de.metas.handlingunits.attribute.strategy.IAttributeStrategy}.
 	 *
 	 * @param onlyHUStatus may be <code>null</code> or empty. Otherwise, only HUs with the given status are updated. However, all HUs are iterated.
@@ -69,6 +68,4 @@ public interface IHUAttributesBL extends ISingletonService
 	boolean isAutomaticallySetLotNumber();
 
 	boolean isAutomaticallySetBestBeforeDate();
-
-	void adjustStorageFromNetWeight(IAttributeSet attributeSet);
 }
