+++ backend/de.metas.handlingunits.base/src/main/java/org/adempiere/mm/attributes/spi/impl/WeightNetAttributeValueCallout.java
@@ -54,7 +54,7 @@ public class WeightNetAttributeValueCallout extends AbstractWeightAttributeValue
 			final Object valueNew)
 	{
 		recalculateWeightGross(attributeSet);
-		huAttributesBL.adjustStorageFromNetWeight(attributeSet);
+		huReceiptScheduleBL.adjustStorageFromNetWeight(attributeSet);
 	}
 
 	/**
